fix: Simplify components tab
