+++ frontend/src/builder.d.ts
@@ -65,7 +65,7 @@ declare type HashString = `#${string}`;
 
 declare type RGBString = `rgb(${number}, ${number}, ${number})`;
 
-declare type LeftSidebarTabOption = "Components" | "Layers";
+declare type LeftSidebarTabOption = "Assets" | "Layers";
 declare type RightSidebarTabOption = "Properties" | "Script" | "Settings";
 
 declare type BuilderMode = "select" | "text" | "container" | "image" | "repeater" | "move";
