+++ frontend/src/components/BuilderLeftPanel.vue
@@ -24,7 +24,7 @@
 			</div>
 			<div class="flex w-full border-gray-200 p-[2px] text-sm dark:border-zinc-800">
 				<button
-					v-for="tab of ['Layers', 'Components']"
+					v-for="tab of ['Layers', 'Assets'] as LeftSidebarTabOption[]"
 					:key="tab"
 					class="mx-3 flex-1 p-2"
 					@click.stop="setActiveTab(tab as LeftSidebarTabOption)"
@@ -36,8 +36,8 @@
 					{{ tab }}
 				</button>
 			</div>
-			<div v-show="store.leftPanelActiveTab === 'Components'">
-				<BuilderComponents class="p-4 pt-3" />
+			<div v-show="store.leftPanelActiveTab === 'Assets'">
+				<BuilderAssets class="p-4 pt-3" />
 			</div>
 			<div v-show="store.leftPanelActiveTab === 'Layers'">
 				<BlockLayers
@@ -59,7 +59,7 @@ import { createResource } from "frappe-ui";
 import { Ref, inject, ref, watch } from "vue";
 import useStore from "../store";
 import BlockLayers from "./BlockLayers.vue";
-import BuilderComponents from "./BuilderComponents.vue";
+import BuilderAssets from "./BuilderAssets.vue";
 import PanelResizer from "./PanelResizer.vue";
 
 import Block from "@/utils/block";
