+++ frontend/src/pages/PageBuilder.vue
@@ -7,6 +7,7 @@
 				class="no-scrollbar absolute bottom-0 left-0 top-[var(--toolbar-height)] z-20 overflow-auto border-r-[1px] bg-white dark:border-gray-800 dark:bg-zinc-900"></BuilderLeftPanel>
 			<BuilderCanvas
 				ref="componentCanvas"
+				:key="store.editingComponent"
 				v-if="store.editingComponent"
 				:block-data="store.getComponentBlock(store.editingComponent)"
 				:canvas-styles="{
